refactor(YSP-840): move integration logic to new module
