+++ web/profiles/custom/yalesites_profile/modules/custom/ys_campus_groups/ys_campus_groups.info.yml
@@ -9,3 +9,4 @@ dependencies:
   - migrate_tools:migrate_tools
   - drupal:key
   - ys_core
+  - ys_integrations
