+++ web/profiles/custom/yalesites_profile/modules/custom/ys_campus_groups/ys_campus_groups.links.menu.yml
@@ -2,6 +2,6 @@ ys_campus_groups.settings:
   title: 'Campus Groups settings'
   description: 'Configure Campus Groups sync settings'
   route_name: ys_campus_groups.settings
-  parent: ys_core.admin_yalesites_integrations
+  parent: ys_integrations.admin_yalesites_integrations
   weight: 0
