+++ web/profiles/custom/yalesites_profile/modules/custom/ys_core/ys_core.links.menu.yml
@@ -54,13 +54,6 @@ ys_core.menu_interface:
   route_name: entity.menu.collection
   title: Manage Menus
   weight: 20
-# Create a new section for integrations
-ys_core.admin_yalesites_integrations:
-  description: "YaleSites integrations with other systems."
-  parent: ys_core.admin_yalesites
-  route_name: ys_core.admin_yalesites_integrations
-  title: "Integrations"
-  weight: 30
 # main menu edit form.
 ys_core.menu_main_edit:
   description: "Manage main navigation"
@@ -158,11 +151,4 @@ ys_core.custom_block_library:
   title: Manage Reusable Blocks
   route_parameters:
     menu: "main"
-  weight: 20
-# Integration settings
-ys_core.integration_settings:
-  description: "Manage integrations exposed to site admins"
-  parent: system.admin_config_services
-  route_name: ys_core.integrations_settings
-  title: "Integration Settings"
   weight: 20
