+++ web/profiles/custom/yalesites_profile/modules/custom/ys_core/ys_core.module
@@ -66,9 +66,6 @@ function ys_core_theme($existing, $type, $theme, $path): array {
         'footer_links_col_2' => [],
       ],
     ],
-    'ys_integrations_block' => [
-      'variables' => ['content' => NULL],
-    ],
   ];
 }
