+++ web/profiles/custom/yalesites_profile/modules/custom/ys_integrations/README.md
@@ -1,0 +1,13 @@
+# ys_integrations Module
+
+## Overview
+The `ys_integrations` module provides tools for managing and configuring third-party integrations in Drupal. It includes a centralized dashboard for site and platform administrators to oversee and control integration settings.
+
+## Installation
+1. Download the module and place it in the `/modules/custom` directory of your Drupal installation.
+2. Enable the module using Drush or through the Drupal admin interface:
+   - Using Drush: `drush en ys_integrations`
+   - Through the admin interface: Navigate to Extend and find `ys_integrations` in the list.
+
+## Contributing
+If you wish to contribute to the `ys_integrations` module, please fork the repository and submit a pull request with your changes.
