+++ web/profiles/custom/yalesites_profile/modules/custom/ys_integrations/src/IntegrationPluginInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\ys_core;
+namespace Drupal\ys_integrations;
 
 /**
  * Defines an interface for integration plugins.
