+++ web/profiles/custom/yalesites_profile/modules/custom/ys_integrations/ys_integrations.info.yml
@@ -1,0 +1,8 @@
+name: 'YS Integrations'
+type: module
+description: 'Provides integrations for YS.'
+core_version_requirement: ^8 || ^9 || ^10
+package: YaleSites
+dependencies:
+  - drupal:user
+  - ys_core
