+++ web/profiles/custom/yalesites_profile/modules/custom/ys_integrations/ys_integrations.links.menu.yml
@@ -1,0 +1,14 @@
+# Create a new section for integrations
+ys_integrations.admin_yalesites_integrations:
+  description: "YaleSites integrations with other systems."
+  parent: ys_core.admin_yalesites
+  route_name: ys_integrations.admin_yalesites_integrations
+  title: "Integrations"
+  weight: 30
+# Integration settings
+ys_integrations.integration_settings:
+  description: "Manage integrations exposed to site admins"
+  parent: system.admin_config_services
+  route_name: ys_integrations.integrations_settings
+  title: "Integration Settings"
+  weight: 20
