+++ web/profiles/custom/yalesites_profile/modules/custom/ys_integrations/ys_integrations.permissions.yml
@@ -1,0 +1,3 @@
+yalesites manage integrations:
+  title: "Manage YaleSite Integrations"
+  description: 'Access the integration settings to expose to YaleSites site admins.'
