+++ web/profiles/custom/yalesites_profile/modules/custom/ys_localist/ys_localist.links.menu.yml
@@ -2,5 +2,5 @@ ys_localist.settings:
   title: 'Localist settings'
   description: 'Configure Localist sync settings'
   route_name: ys_localist.settings
-  parent: ys_core.admin_yalesites_integrations
+  parent: ys_integrations.admin_yalesites_integrations
   weight: 0
